+++ test/perf/frame_and_tap_bench_test.dart
@@ -173,6 +173,7 @@ class _NullSink implements CaptureSink {
     required String name,
     Map<String, Object?> props = const {},
     String? screen,
+    int? tsMs,
   }) {}
 }